新增部门，和更改主键类型为Integer

+++ authority-dal/src/main/java/com/chenjianwen/dal/authority/SysAclMapper.java
@@ -1,18 +1,18 @@
 package com.chenjianwen.dal.authority;
 
-import com.chenjianwen.api.authority.SysAcl;
+import com.chenjianwen.api.authority.po.SysAcl;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SysAclMapper {
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(SysAcl record);
 
     int insertSelective(SysAcl record);
 
-    SysAcl selectByPrimaryKey(Long id);
+    SysAcl selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysAcl record);
