+++ authority-dal/src/main/java/com/chenjianwen/dal/authority/SysAclModuleMapper.java
@@ -1,17 +1,17 @@
 package com.chenjianwen.dal.authority;
 
-import com.chenjianwen.api.authority.SysAclModule;
+import com.chenjianwen.api.authority.po.SysAclModule;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SysAclModuleMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(SysAclModule record);
 
     int insertSelective(SysAclModule record);
 
-    SysAclModule selectByPrimaryKey(Long id);
+    SysAclModule selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysAclModule record);
