+++ authority-dal/src/main/java/com/chenjianwen/dal/authority/SysRoleMapper.java
@@ -1,17 +1,17 @@
 package com.chenjianwen.dal.authority;
 
-import com.chenjianwen.api.authority.SysRole;
+import com.chenjianwen.api.authority.po.SysRole;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SysRoleMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(SysRole record);
 
     int insertSelective(SysRole record);
 
-    SysRole selectByPrimaryKey(Long id);
+    SysRole selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysRole record);
